TransposeSinking Rafactoring part2: move the transformations to a separate folder, align namespaces
