+++ src/common/transformations/include/transformations/transpose_sinking/transpose_sinking_fuse.hpp
