+++ src/common/transformations/src/transformations/transpose_sinking/transpose_sinking_split.cpp
@@ -2,19 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "transformations/common_optimizations/transpose_sinking_split.hpp"
-
-#include <openvino/pass/pattern/op/or.hpp>
-#include <transformations/utils/utils.hpp>
-#include <utility>
+#include "transformations/transpose_sinking/transpose_sinking_split.hpp"
 
 #include "itt.hpp"
 #include "openvino/op/util/op_types.hpp"
 #include "openvino/opsets/opset10.hpp"
 #include "openvino/pass/pattern/op/label.hpp"
 #include "openvino/pass/pattern/op/wrap_type.hpp"
-#include "transformations/common_optimizations/transpose_sinking_utils.hpp"
+#include "transformations/transpose_sinking/transpose_sinking_utils.hpp"
 #include "transformations/rt_info/transpose_sinking_attr.hpp"
+#include "transformations/utils/utils.hpp"
 
 using namespace ov::pass::pattern;
 using namespace ov;
