+++ src/common/transformations/src/transformations/transpose_sinking/transpose_sinking_utils.cpp
@@ -2,15 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "transformations/common_optimizations/transpose_sinking_utils.hpp"
-
-#include <transformations/utils/utils.hpp>
+#include "transformations/transpose_sinking/transpose_sinking_utils.hpp"
 
 #include "itt.hpp"
 #include "openvino/op/util/op_types.hpp"
 #include "openvino/opsets/opset10.hpp"
 #include "openvino/util/common_util.hpp"
 #include "transformations/rt_info/transpose_sinking_attr.hpp"
+#include "transformations/utils/utils.hpp"
 
 namespace transpose_sinking {
