+++ src/common/transformations/tests/transpose_sinking/transpose_sinking_test_utils.hpp
@@ -4,15 +4,16 @@
 
 #pragma once
 
-#include <openvino/frontend/manager.hpp>
-#include <openvino/opsets/opset10.hpp>
-#include <openvino/pass/manager.hpp>
+#include "openvino/frontend/manager.hpp"
+#include "openvino/opsets/opset10.hpp"
+#include "openvino/pass/manager.hpp"
 
 #include "common_test_utils/ngraph_test_utils.hpp"
 #include "gtest/gtest.h"
 
 namespace transpose_sinking {
 namespace testing {
+namespace utils {
 
 using NodePtr = std::shared_ptr<ov::Node>;
 
@@ -67,5 +68,6 @@ std::shared_ptr<ov::Node> constant(ov::element::Type el_type, const ov::Shape& s
     return ov::opset10::Constant::create<T>(el_type, shape, value);
 }
 
+}  // utils
 }  // namespace testing
 }  // namespace transpose_sinking
