+++ src/frontends/tensorflow/src/frontend.cpp
@@ -21,7 +21,7 @@
 #include "so_extension.hpp"
 #include "tf_framework_node.hpp"
 #include "transformations/common_optimizations/reverse_shape_and_type_infer.hpp"
-#include "transformations/common_optimizations/transpose_sinking_general.hpp"
+#include "transformations/transpose_sinking/transpose_sinking_general.hpp"
 #include "translate_session.hpp"
 #include "utils.hpp"
